+++ rest/src/decorators/httpClientErrorForceCustomMessage.decorator.ts
@@ -1,12 +1,14 @@
 import type {RESTClientBase, RestMethodMiddlewares} from '@anglr/rest';
+import {isPresent} from '@jscrpt/common';
 
 import {RestCatchHttpClientError} from '../interfaces';
 
 /**
  * Sets http client error to force custom message displaying
  * @param indication - Indication whether force custom message displaying, defaults to true
+ * @param statusCode - If specified, force custom message display will be set for specific http status code
  */
-export function HttpClientErrorForceCustomMessage(indication: boolean = true)
+export function HttpClientErrorForceCustomMessage(indication: boolean = true, statusCode?: number)
 {
     return function<TDecorated>(_target: RESTClientBase, _propertyKey: string, descriptor: RestCatchHttpClientError &
                                                                                            RestMethodMiddlewares |
@@ -14,7 +16,16 @@ export function HttpClientErrorForceCustomMessage(indication: boolean = true)
     {
         const descr = descriptor as RestCatchHttpClientError & RestMethodMiddlewares;
 
-        descr.forceCustomMessageDisplay = indication;
+        if(isPresent(statusCode))
+        {
+            descr.configs ??= {};
+            descr.configs[statusCode] ??= {};
+            descr.configs[statusCode].forceCustomMessageDisplay = indication;
+        }
+        else
+        {
+            descr.forceCustomMessageDisplay = indication;
+        }
 
         return descr as TDecorated;
     };