+++ src/misc/providers.ts
@@ -1,9 +1,9 @@
 import {ClassProvider, EnvironmentProviders, FactoryProvider, Provider, Type, ValueProvider, makeEnvironmentProviders, inject} from '@angular/core';
 import {extend} from '@jscrpt/common/extend';
 
-import {ANGLR_EXCEPTION_EXTENDERS, HTTP_CLIENT_ERROR_HANDLERS, HTTP_CLIENT_ERROR_MESSAGES, HTTP_CLIENT_ERROR_RESPONSE_MAPPER, HTTP_CLIENT_VALIDATION_ERROR_RESPONSE_MAPPER, INTERNAL_SERVER_ERROR_RENDERER} from './tokens';
+import {ANGLR_EXCEPTION_EXTENDERS, HTTP_CLIENT_ERROR_HANDLERS, HTTP_CLIENT_ERROR_CONFIGS, HTTP_CLIENT_ERROR_RESPONSE_MAPPER, HTTP_CLIENT_VALIDATION_ERROR_RESPONSE_MAPPER, INTERNAL_SERVER_ERROR_RENDERER} from './tokens';
 import {AnglrExceptionExtender, InternalServerErrorRenderer} from '../interfaces';
-import {HttpClientErrorHandlers, HttpClientErrorMessages, HttpClientErrorResponseMapper, HttpClientValidationErrorResponseMapper} from './types';
+import {HttpClientErrorConfigs, HttpClientErrorHandlers, HttpClientErrorResponseMapper, HttpClientValidationErrorResponseMapper} from './types';
 
 /**
  * Provides anglr exception extends functions
@@ -38,18 +38,18 @@ export function provideInternalServerErrorRenderer(renderer: Type<InternalServer
 }
 
 /**
- * Provides http client error messages and merges them with existing provided messages
- * @param messages - Object storing messages to be provided
+ * Provides http client error configs and merges them with existing provided configs
+ * @param configs - Object storing configs to be provided
  */
-export function provideHttpClientErrorMessages(messages: HttpClientErrorMessages): Provider
+export function provideHttpClientErrorConfigs(configs: HttpClientErrorConfigs): Provider
 {
     return <FactoryProvider> {
-        provide: HTTP_CLIENT_ERROR_MESSAGES,
+        provide: HTTP_CLIENT_ERROR_CONFIGS,
         useFactory: () =>
         {
-            const existingMessages = inject(HTTP_CLIENT_ERROR_MESSAGES, {skipSelf: true, optional: true});
+            const existingConfigs = inject(HTTP_CLIENT_ERROR_CONFIGS, {skipSelf: true, optional: true});
 
-            return extend({}, existingMessages, messages);
+            return extend({}, existingConfigs, configs);
         },
     };
 }