+++ src/rxjsOperators/catchHttpClientError.operator.ts
@@ -5,8 +5,8 @@ import {Observable, OperatorFunction, catchError, throwError} from 'rxjs';
 import {CatchHttpClientErrorOptions} from '../interfaces';
 import {handleHttpClientErrors} from '../errorHandlers';
 import {HttpClientError} from '../misc/classes/httpClientError';
-import {HttpClientErrorHandlers, HttpClientErrorMessages} from '../misc/types';
-import {HTTP_CLIENT_ERROR_HANDLERS, HTTP_CLIENT_ERROR_MESSAGES} from '../misc/tokens';
+import {HttpClientErrorConfigs, HttpClientErrorHandlers} from '../misc/types';
+import {HTTP_CLIENT_ERROR_HANDLERS, HTTP_CLIENT_ERROR_CONFIGS} from '../misc/tokens';
 
 /**
  * Catches http client errors and handles them according provided options
@@ -15,7 +15,7 @@ import {HTTP_CLIENT_ERROR_HANDLERS, HTTP_CLIENT_ERROR_MESSAGES} from '../misc/to
 export function catchHttpClientError<TIn>(options?: CatchHttpClientErrorOptions): OperatorFunction<TIn, TIn|HttpClientError>
 {
     const injector: Injector = options?.injector ?? inject(Injector);
-    const messages: HttpClientErrorMessages = extend({}, injector.get(HTTP_CLIENT_ERROR_MESSAGES, {}, {optional: true}), options?.messages);
+    const configs: HttpClientErrorConfigs = extend({}, injector.get(HTTP_CLIENT_ERROR_CONFIGS, {}, {optional: true}), options?.configs);
     const handlers: HttpClientErrorHandlers = extend({}, injector.get(HTTP_CLIENT_ERROR_HANDLERS, {}, {optional: true}), options?.handlers);
 
     return source =>
@@ -38,7 +38,7 @@ export function catchHttpClientError<TIn>(options?: CatchHttpClientErrorOptions)
                         {
                             ...options,
                             injector,
-                            messages,
+                            configs,
                             handlers,
                         });